Update tutorials
Utilizes thew new macros introduced in the previous commit.
