+++ Sources/SwiftGodot/SwiftGodot.docc/Resources/Tutorials/Sources/MainLevel-multiscript-6.swift
@@ -7,17 +7,21 @@
 
 import Foundation
 import SwiftGodot
+import SwiftGodotMacros
 
 @Godot
 class MainLevel: Node2D {
-    var player: PlayerController?
-    var spawnpoint: Node2D?
-    var teleportArea: Area2D?
+    @SceneTree(path: "CharacterBody2D") var player: PlayerController?
+    @SceneTree(path: "Spawnpoint") var spawnpoint: Node2D?
+    @SceneTree(path: "Telepoint") var teleportArea: Area2D?
 
     override func _ready() {
-        self.player = getNodeOrNull(path: "CharacterBody2D") as? PlayerController
-        self.teleportArea = getNodeOrNull(path: "Telepoint") as? Area2D
-        self.spawnpoint = getNodeOrNull(path: "Spawnpoint") as? Node2D
+        teleportArea?.bodyEntered.connect { [self] enteredBody in
+            if enteredBody.isClass("\(PlayerController.self)") {
+                teleportPlayerToTop()
+            }
+        }
+
         super._ready()
     }
