+++ Sources/SwiftGodot/SwiftGodot.docc/Resources/Tutorials/Sources/MainLevel-multiscript-init.swift
@@ -7,3 +7,4 @@
 
 import Foundation
 import SwiftGodot
+import SwiftGodotMacros
